Finished base refactor
