+++ common/Cargo.toml
@@ -9,6 +9,7 @@ cortex-m = "0.7.3"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 alloc-traits = "0.1.1"
 ordered-float = { version = "2.0", default-features = false, features = ["serde"] }
+static-alloc = "0.2.3"
 
 [features]
 std = []
